edhoc: stop truncating received message size to 8 bits
The received message size was cast to uint8_t before being stored in
buffers.rx_sz, which is a uint16_t. This truncated the size of any
EDHOC message or block larger than 255 bytes. Drop the cast.

+++ os/net/security/edhoc/edhoc-client.c
@@ -268,7 +268,7 @@ client_response_handler(coap_callback_request_state_t *callback_state)
   client_block2_handler(callback_state->state.response,
                         rx_ptr, &rx_sz, EDHOC_MAX_PAYLOAD_LEN);
   if(!callback_state->state.more) {
-    edhoc_ctx->buffers.rx_sz = (uint8_t)rx_sz;
+    edhoc_ctx->buffers.rx_sz = rx_sz;
     edhoc_state.val = EVT_BLOCKING;
     process_post(PROCESS_BROADCAST, edhoc_event, &edhoc_state);
   }
